added image folder

+++ lib/ui/extension/media_query.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 
 extension BuildContextMediaQuery on BuildContext {
-  double widthPercent(int percentage) =>
+  double widthPercent(double percentage) =>
       MediaQuery.of(this).size.width * percentage * 0.01;
 
-  double heightPercent(int percentage) =>
+  double heightPercent(double percentage) =>
       MediaQuery.of(this).size.height * percentage * 0.01;
 }
